Removed unneeded additional args to `comment`
