+++ layers/Schema/packages/comments/src/ModuleProcessors/SingleCommentFilterInputContainerModuleProcessor.php
@@ -1,0 +1,69 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoPSchema\Comments\ModuleProcessors;
+
+use PoP\ComponentModel\FilterInput\FilterInputHelper;
+use PoP\ComponentModel\Schema\SchemaTypeModifiers;
+use PoPSchema\Comments\ModuleProcessors\FormInputs\FilterInputModuleProcessor;
+use PoPSchema\SchemaCommons\ModuleProcessors\AbstractFilterInputContainerModuleProcessor;
+use PoPSchema\SchemaCommons\ModuleProcessors\FormInputs\CommonFilterInputModuleProcessor;
+
+class SingleCommentFilterInputContainerModuleProcessor extends AbstractFilterInputContainerModuleProcessor
+{
+    public const HOOK_FILTER_INPUTS = __CLASS__ . ':filter-inputs';
+
+    public const MODULE_FILTERINPUTCONTAINER_COMMENT_STATUS = 'filterinputcontainer-comment-status';
+    public const MODULE_FILTERINPUTCONTAINER_COMMENT_BY_ID_STATUS = 'filterinputcontainer-comment-by-id-status';
+    
+    public function getModulesToProcess(): array
+    {
+        return array(
+            [self::class, self::MODULE_FILTERINPUTCONTAINER_COMMENT_STATUS],
+            [self::class, self::MODULE_FILTERINPUTCONTAINER_COMMENT_BY_ID_STATUS],
+        );
+    }
+
+    public function getFilterInputModules(array $module): array
+    {
+        return match ((string)$module[1]) {
+            self::MODULE_FILTERINPUTCONTAINER_COMMENT_STATUS => [
+                [FilterInputModuleProcessor::class, FilterInputModuleProcessor::MODULE_FILTERINPUT_COMMENT_STATUS],
+            ],
+            self::MODULE_FILTERINPUTCONTAINER_COMMENT_BY_ID_STATUS => [
+                [CommonFilterInputModuleProcessor::class, CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_ID],
+                [FilterInputModuleProcessor::class, FilterInputModuleProcessor::MODULE_FILTERINPUT_COMMENT_STATUS],
+            ],
+            default => [],
+        };
+    }
+
+    public function getFieldFilterInputTypeModifiers(array $module, string $fieldArgName): int
+    {
+        $fieldFilterInputTypeModifiers = parent::getFieldFilterInputTypeModifiers($module, $fieldArgName);
+        switch ($module[1]) {
+            case self::MODULE_FILTERINPUTCONTAINER_COMMENT_BY_ID_STATUS:
+                $idFilterInputName = FilterInputHelper::getFilterInputName([
+                    CommonFilterInputModuleProcessor::class,
+                    CommonFilterInputModuleProcessor::MODULE_FILTERINPUT_ID
+                ]);
+                if ($fieldArgName === $idFilterInputName) {
+                    return $fieldFilterInputTypeModifiers | SchemaTypeModifiers::MANDATORY;
+                }
+                break;
+        }
+        return $fieldFilterInputTypeModifiers;
+    }
+
+    /**
+     * @return string[]
+     */
+    protected function getFilterInputHookNames(): array
+    {
+        return [
+            ...parent::getFilterInputHookNames(),
+            self::HOOK_FILTER_INPUTS,
+        ];
+    }
+}
